requirements-dev.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
boto3==1.36.22
botocore==1.36.22
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.6
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-filter==25.1
django-model-utils==5.0.0
django-storages==1.14.5
djangorestframework==3.15.2
jmespath==1.0.1
jsonfield==3.1.0
kombu==5.4.2
mysqlclient==2.2.6
prompt_toolkit==3.0.50
PyMySQL==1.0.2
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
swapper==1.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
